# apidocs/swagger.yaml
basePath: /
definitions:
  github_com_stockflow_server_internal_response.Envelope:
    properties:
      code:
        description: 成功 0（数字）；失败为模块命名空间错误码字符串（COMMON_* 等）
      data: {}
      details:
        description: 校验失败必带（api.md §4）
      message:
        type: string
      request_id:
        type: string
    type: object
...
info:
  contact: {}
  description: StockFlow（库流智能仓储管理系统）后端 API。统一信封 {code,message,data,request_id,details}（api.md
    §2）。
  title: StockFlow API
  version: "1.0"
paths:
...
  /api/customers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/customers
      tags:
      - 基础资料
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/customers
      tags:
      - 基础资料
  /api/customers/{id}:
    delete:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: DELETE /api/customers/:id
      tags:
      - 基础资料
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/customers/:id
      tags:
      - 基础资料
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/customers/:id
      tags:
      - 基础资料
  /api/customers/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/customers/:id/status
      tags:
      - 基础资料
...
  /api/product-categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/product-categories
      tags:
      - 基础资料
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/product-categories
      tags:
      - 基础资料
  /api/product-categories/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/product-categories/:id
      tags:
      - 基础资料
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/product-categories/:id
      tags:
      - 基础资料
  /api/product-categories/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/product-categories/:id/status
      tags:
      - 基础资料
  /api/products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/products
      tags:
      - 基础资料
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/products
      tags:
      - 基础资料
  /api/products/{id}:
    delete:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: DELETE /api/products/:id
      tags:
      - 基础资料
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/products/:id
      tags:
      - 基础资料
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/products/:id
      tags:
      - 基础资料
  /api/products/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/products/:id/status
      tags:
      - 基础资料
...
  /api/skus:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/skus
      tags:
      - 基础资料
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/skus
      tags:
      - 基础资料
...
  /api/skus/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/skus/:id/status
      tags:
      - 基础资料
  /api/suppliers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/suppliers
      tags:
      - 基础资料
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/suppliers
      tags:
      - 基础资料
...
securityDefinitions:
  BearerAuth:
    description: 值格式 "Bearer <access_token>"（api.md §6.1）。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"